master_upsc_processing/requirements.txt
mistralai>=0.0.10
pydantic>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
uuid>=1.30
PyPDF2>=3.0.0
openai>=1.0.0
tqdm>=4.65.0
supabase>=2.0.0 